+++ tests/PyroTests/test_core.py
@@ -295,6 +295,8 @@ class CoreTests(unittest.TestCase):
         p1._pyroOneway = set("def")
         p1._pyroMethods = set("ghi")
         p1._pyroHmacKey = b"secret"
+        p1._pyroHandshake = "apples"
+        p1._pyroStuff = [1, 2, 3]
         p2 = copy.copy(p1)
         self.assertEqual(p1, p2)
         self.assertEqual(p1._pyroUri, p2._pyroUri)
@@ -303,6 +305,8 @@ class CoreTests(unittest.TestCase):
         self.assertEqual(p1._pyroAttrs, p2._pyroAttrs)
         self.assertEqual(p1._pyroTimeout, p2._pyroTimeout)
         self.assertEqual(p1._pyroHmacKey, p2._pyroHmacKey)
+        self.assertEqual(p1._pyroHandshake, p2._pyroHandshake)
+        self.assertEqual(p1._pyroStuff, p2._pyroStuff)
 
     def testProxyOffline(self):
         # only offline stuff here.